refactor(loop): split tick.py under the 500-LOC module-health gate (#1054)
`src/teatree/loop/tick.py` was 517 LOC on `origin/main` and tripped
`scripts/hooks/check_module_health.py` (max 500). The gate has no
`--no-verify` or `relax:` escape (see `t3:rules` § "No commit-type
bypass for the quality gates"), so this was a structural blocker
for any feature branch — #1049 in particular.

Behaviour-preserving split into three sibling modules by concern:

* `tick_freshness.py` — repo-currency / statusline-header freshness.
    Contains `_repo_freshness`, `_repos_from_toml`,
    `_canonical_overlay_names`, `_collect_repo_freshness`,
    `_write_tick_meta`.
* `tick_recovery.py` — boot/tick recovery and post-dispatch
    side-effects. Contains `_reap_stale_task_claims`,
    `_record_dashboard_actions`, `_persist_agent_dispatches`,
    `_execute_mechanical`.
* `tick_jobs.py` — scanner-job construction. Contains `_ScannerJob`,
    `_jobs_for_backend_hosts`, `_gitlab_approvals_enabled`,
    `_run_job`, `_user_slack_id_for_overlay`,
    `_user_identity_aliases_for_overlay`, `build_default_jobs`,
    `build_default_scanners`.

`tick.py` keeps `TickReport`, `TickRequest`, `run_tick` and
re-exports every moved name (`__all__` and module-level imports) so
no downstream import breaks. Function bodies are unchanged — pure
move.

Post-split LOC (per `check_module_health.py` counting):

* `tick.py`            153
* `tick_jobs.py`       265
* `tick_recovery.py`    91
* `tick_freshness.py`  106

Test-side change: a few `monkeypatch` targets moved from
`teatree.loop.tick.load_config`/`discover_overlays` to
`teatree.loop.tick_jobs.*` / `teatree.loop.tick_recovery.*` so the
patch hits the binding that the moved functions actually look up.
The imported helpers in `teatree.loop.tick` still resolve via
re-export, so production callers and other test imports are
untouched.

BLUEPRINT § "loop/ topology" updated to list the new sibling
modules.

All 492 tests in `tests/teatree_loop/` and
`tests/teatree_core/test_loop_tick_command.py` pass green.
